move zero state from amount page to send page
