+++ src/pages/send/amount/amount.html
@@ -15,24 +15,7 @@
 
 <ion-content no-bounce>
 
-  <div class="prompt-user-fullscreen-container" *ngIf="shouldShowZeroState()">
-    <div class="prompt-user">
-      <div class="title-icon larger-icon">
-        <img src="assets/img/ghost-straight-face.svg" />
-      </div>
-      <div class="title-info">
-        <span translate>Mmm...</span>
-      </div>
-      <div class="subtitle-info">
-        <span translate>It appears that you do not have any money to send.</span>
-      </div>
-      <button ion-button class="button-standard" (click)="goToReceive()">
-        {{'Add funds' | translate}}
-      </button>
-    </div>
-  </div>
-
-  <div *ngIf="!shouldShowZeroState()" class="amount-wrapper">
+  <div class="amount-wrapper">
 
     <div class="amount-header-margin">
       <div class="amount-header">
